+++ testcenterservice/src/main/java/com/zoctan/api/controller/EnviromentController.java
@@ -5,7 +5,11 @@ import com.github.pagehelper.PageInfo;
 import com.zoctan.api.core.response.Result;
 import com.zoctan.api.core.response.ResultGenerator;
 import com.zoctan.api.entity.Enviroment;
+import com.zoctan.api.entity.Envmachine;
+import com.zoctan.api.entity.Macdepunit;
 import com.zoctan.api.service.EnviromentService;
+import com.zoctan.api.service.EnvmachineService;
+import com.zoctan.api.service.MacdepunitService;
 import org.springframework.web.bind.annotation.*;
 import tk.mybatis.mapper.entity.Condition;
 
@@ -22,6 +26,8 @@ import java.util.Map;
 public class EnviromentController {
     @Resource
     private EnviromentService enviromentService;
+    @Resource
+    private MacdepunitService macdepunitService;
 
     @PostMapping
     public Result add(@RequestBody Enviroment enviroment) {
@@ -39,8 +45,18 @@ public class EnviromentController {
 
     @DeleteMapping("/{id}")
     public Result delete(@PathVariable Long id) {
-        enviromentService.deleteById(id);
-        return ResultGenerator.genOkResult();
+        //增加判断此服务器当前是否有环境在使用
+        List<Macdepunit>macdepunitList= macdepunitService.findenviromentbyenvid(id);
+        if(macdepunitList.size()>0)
+        {
+            return ResultGenerator.genFailedResult("当前环境在环境部署中还在使用，无法删除！");
+        }
+        else
+        {
+            enviromentService.deleteById(id);
+            return ResultGenerator.genOkResult();
+        }
+
     }
 
     @PatchMapping
